Added the actual files
These are the files involved in the creation of the Speck.dll.  I wrote
the .asm file as source code and the .def file for the linker to export
the right functions.  The .dll is the final product, and the other files
were created by the assembler or linker in the build process.